New: Plugins. Putting together rules, schema and pipeline.
These plugins are all transacted in to
the meta db and used whenever query or
index is called. Placing them together
Makes it easier to write pieces of logic
that should be together and removes the
responsibility on the developer to add
them in to the right function calls.
Instead one installs them once, and they're
included where they should be.